# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/academies": {
      "post": {
        "operationId": "academies_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin-only. There are roughly fifty academies in Uzbekistan, so they are\nonboarded by the platform team rather than self-registered — see\nAcademiesService.register for the reasoning.",
        "tags": ["academies"]
      },
      "get": {
        "operationId": "academies_listPublic",
        "parameters": [
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/mine": {
      "get": {
        "operationId": "academies_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The academy the caller manages, or null.\n\nDeclared before `:id` — Nest matches routes in declaration order, so putting\nthis after would make `/academies/mine` resolve as an academy whose id is the\nliteral string \"mine\".",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}": {
      "get": {
        "operationId": "academies_getPublicProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      },
      "patch": {
        "operationId": "academies_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manager edits their own; an admin may correct any (§1.10).",
        "tags": ["academies"]
      },
      "delete": {
        "operationId": "academies_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archives an academy (admin only). Sets status REJECTED rather than deleting —\na hard delete would cascade through trials, applications and recommendation\nhistory. See AcademiesService.archive.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/relation": {
      "get": {
        "operationId": "academies_relation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What the caller is to this academy — manager, staff, endorsed, or a player it\naccepted at a trial. Drives the \"My academy\" badge.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/manager": {
      "patch": {
        "operationId": "academies_setManager",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetManagerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assigns or replaces the academy's single manager (admin only). Either names an\nexisting user or mints an account, returning its one-time credentials once.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/manager/reset-password": {
      "post": {
        "operationId": "academies_resetManagerPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issues the manager a fresh one-time password (admin only).",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/admin/all": {
      "get": {
        "operationId": "academies_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every academy including pending/archived — admin console list.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/coaches": {
      "post": {
        "operationId": "academies_createCoach",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCoachDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a coach: an existing account, or a new one minted with credentials the\nmanager hands over — the same two paths an admin has for a manager.\n\nCredentials come back exactly once and are never retrievable again.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/groups": {
      "get": {
        "operationId": "academies_listGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      },
      "post": {
        "operationId": "academies_createGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/groups/mine": {
      "get": {
        "operationId": "academies_myGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A coach's own groups — declared before `:id/groups/:groupId`.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/groups/{groupId}": {
      "get": {
        "operationId": "academies_getGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      },
      "patch": {
        "operationId": "academies_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      },
      "delete": {
        "operationId": "academies_deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deleting a group returns its people to the reserve; it removes nobody.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/groups/move": {
      "post": {
        "operationId": "academies_moveMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveMembersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move members into a group, or back to the reserve by omitting `groupId`.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/candidates": {
      "get": {
        "operationId": "academies_joinCandidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Declared here rather than read separately: the global ValidationPipe runs\n`forbidNonWhitelisted` over the whole query object, so a parameter missing\nfrom this DTO is a 400 even when the handler binds it by name.",
            "schema": {
              "enum": ["PLAYER", "COACH", "SCOUT"],
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accounts this academy could add for a role — declared before `:id/transfers`.\n\nPaged and searchable: the picker on the squad screen is a window onto every\naccount holding that role, and there are far more of those than fit a list.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/{id}/invitations": {
      "post": {
        "operationId": "academies_invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      },
      "get": {
        "operationId": "academies_listAcademyInvitations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What this academy has asked of people, answered or not.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/invitations/mine": {
      "get": {
        "operationId": "academies_listMyInvitations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invitations addressed to me — declared before `:id` cannot catch it.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/invitations/{invitationId}/accept": {
      "post": {
        "operationId": "academies_acceptInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/invitations/{invitationId}/reject": {
      "post": {
        "operationId": "academies_rejectInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/invitations/{invitationId}/cancel": {
      "post": {
        "operationId": "academies_cancelInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The academy withdrawing a question nobody has answered yet.",
        "tags": ["academies"]
      }
    },
...
    "/api/v1/academies/transfers/{transferId}/approve": {
      "post": {
        "operationId": "academies_approveTransfer",
        "parameters": [
          {
            "name": "transferId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The receiving academy accepts — the member lands in their reserve.",
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/transfers/{transferId}/reject": {
      "post": {
        "operationId": "academies_rejectTransfer",
        "parameters": [
          {
            "name": "transferId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["academies"]
      }
    },
    "/api/v1/academies/transfers/{transferId}/cancel": {
      "post": {
        "operationId": "academies_cancelTransfer",
        "parameters": [
          {
            "name": "transferId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The offering academy withdraws before the other side answers.",
        "tags": ["academies"]
      }
    },
...
  },
  "info": {
    "title": "FotSpot API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` returned by any `/auth/*` login endpoint."
      }
    },
    "schemas": {
...
      "CreateAcademyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "managerUserId": {
            "type": "string"
          },
          "newManager": {
            "$ref": "#/components/schemas/NewManagerDto"
          }
        },
        "required": ["name"]
      },
      "SetManagerDto": {
        "type": "object",
        "properties": {
          "managerUserId": {
            "type": "string"
          },
          "newManager": {
            "$ref": "#/components/schemas/NewManagerDto"
          }
        }
      },
...
      "CreateCoachDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "newCoach": {
            "$ref": "#/components/schemas/NewManagerDto"
          },
          "bio": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
    }
  },
...
}